requirements.txt
Flask==2.3.3
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
PyJWT==2.8.0
python-dotenv==1.0.0
